requirements.txt
django
psycopg2
djangorestframework
channels
django-rest-knox
django-cors-headers
pytest-django
dj-rest-auth
django-allauth
channels-redis
dj-database-url
requests
pyjwt
django-environ